fix: Custom Discord activity name not working
This was due to me testing on Vesktop, which uses arRPC (an alternate,
open source implementation of the receiving end of Discord's Rich
Presence). Its implementation respects applications' requests to show a
custom name - the official Discord client doesn't. This is fixed by
requesting Discord to display the details field as the activity name
instead.